refactor: move snackbar from components to utils
