+++ src/app/manage/[slug]/(dashboard)/(forms)/Details.tsx
@@ -8,7 +8,7 @@ import type z from 'zod';
 import ClubTagAutocomplete from '@src/components/club/ClubTagAutocomplete';
 import Panel, { PanelSkeleton } from '@src/components/common/Panel';
 import Confirmation from '@src/components/Confirmation';
-import { setSnackbar, SnackbarPresets } from '@src/components/global/Snackbar';
+import { setSnackbar, SnackbarPresets } from 'src/utils/snackbar';
 import { ClubSchoolEdit } from '@src/components/manage/form/ClubSchoolEdit';
 import FormImage from '@src/components/manage/form/FormImage';
 import { SelectClub } from '@src/server/db/models';
