+++ src/components/events/directory/filter/EventDirectoryGrid.tsx
@@ -8,7 +8,7 @@ import {
 import { AnimatePresence, motion } from 'framer-motion';
 import { useEffect } from 'react';
 import EventCard, { EventCardVariants } from '@src/components/events/EventCard';
-import { setSnackbar, SnackbarPresets } from '@src/components/global/Snackbar';
+import { setSnackbar, SnackbarPresets } from 'src/utils/snackbar';
 import { useTRPC } from '@src/trpc/react';
 import { RouterOutputs } from '@src/trpc/shared';
 import { EventFiltersSchema } from '@src/utils/eventFilter';
