+++ src/components/settings/forms/UserInfo.tsx
@@ -7,7 +7,7 @@ import { DatePicker } from '@mui/x-date-pickers/DatePicker';
 import { useMutation } from '@tanstack/react-query';
 import { useState } from 'react';
 import Panel from '@src/components/common/Panel';
-import { setSnackbar, SnackbarPresets } from '@src/components/global/Snackbar';
+import { setSnackbar, SnackbarPresets } from 'src/utils/snackbar';
 import { majors, minors } from '@src/constants/utdDegrees';
 import { SelectUserMetadataWithClubs } from '@src/server/db/models';
 import { studentClassificationEnum } from '@src/server/db/schema/users';
